Update ReadMe.me links
The change to Jupyter Books format broke links in the readme, so I am now fixing these direct links. In the future these will probably be supplanted by having users directly access the GitHub Pages site for the book, but for now they provide a useful way for users to navigate the notebooks.